pyproject.toml
[project]
name = "izinscope"
version = "0.3.0"
description = "**izinscope** est un outil en ligne de commande permettant de vérifier si des adresses IP ou des domaines se trouvent dans un scope défini (IP, CIDR, domaines résolus). Il facilite l’identification des ressources “in-scope” et peut générer des rapports TXT ou CSV."
authors = [
    {name = "Mahafaly Randriamiarisoa"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dnspython (>=2.7.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
izinscope = "izinscope.__init__:main"
[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"

